Implemented file downloading via streams

+++ PDFUpload/Controllers/PrintPDF.cs
@@ -5,42 +5,50 @@ using System.Text.Json;
 namespace PDFUpload.Controllers
 {
     [ApiController]
-    [Route("/api/printpdf")]
-    public class WeatherForecastController : ControllerBase
+    [Route("[controller]")]
+    public class PrintPdfController : ControllerBase
     {
-        private readonly ILogger<WeatherForecastController> _logger;
+        private readonly ILogger<PrintPdfController> _logger;
 
-        public WeatherForecastController(ILogger<WeatherForecastController> logger)
+        public PrintPdfController(ILogger<PrintPdfController> logger)
         {
             _logger = logger;
         }
 
         [HttpPost]
-        public async Task<ActionResult<string>> PrintPdf(IFormFile pdf,IFormFile json,string destination_filename)
+        public async Task<ActionResult> PrintPdf(IFormFile pdf, IFormFile json, string destination_filename)
         {
             var currentRoute = System.IO.Directory.GetCurrentDirectory();
             string route = Path.Combine(currentRoute, "files");
-            var pdf_new = Path.Combine(route,$"{Guid.NewGuid()}{Path.GetExtension(pdf.FileName)}");
+            var pdf_new = Path.Combine(route, $"{Guid.NewGuid()}{Path.GetExtension(pdf.FileName)}");
             var json_new = Path.Combine(route, $"{Guid.NewGuid()}{Path.GetExtension(json.FileName)}");
-            
+
 
             SaveFile saveFile = new SaveFile();
-            saveFile.save(pdf,pdf_new);
-            saveFile.save(json,json_new);
+            await saveFile.save(pdf, pdf_new);
+            await saveFile.save(json, json_new);
             MergeDocument sourcepdf = new MergeDocument(@pdf_new);
             var file = System.IO.File.ReadAllText(@json_new);
             var data = JsonSerializer.Deserialize<Dictionary<string, string>>(file);
 
-            foreach (var (key, value) in data)
+            var bytes = await System.IO.File.ReadAllBytesAsync(pdf_new);
+            var returnValue = File(bytes, "application/pdf", destination_filename + ".pdf");
+
+            if (data is not null)
             {
-                sourcepdf.Form.Fields[key].Value = value;
+                foreach (var (key, value) in data)
+                {
+                    sourcepdf.Form.Fields[key].Value = value;
+                }
+                string destpdf = Path.Combine(route, destination_filename + ".pdf");
+                sourcepdf.Draw(@destpdf);
+                bytes = await System.IO.File.ReadAllBytesAsync(destpdf);
+                returnValue = File(bytes, "application/pdf", Path.GetFileName(destpdf));
+                System.IO.File.Delete(@pdf_new);
+                System.IO.File.Delete(@json_new);
+                System.IO.File.Delete(destpdf);
             }
-            //sourcepdf.Form.Fields["CheckBox1"].Value = "1";
-            string destpdf = Path.Combine(route, destination_filename + ".pdf");
-            sourcepdf.Draw(@destpdf);
-            System.IO.File.Delete(@pdf_new);
-            System.IO.File.Delete(@json_new);
-            return "Success";
+            return returnValue;
 
         }
     }
